Переносит View информации и предложений точки в саму View точки
